DZ2/src/user_service/requirements.txt
fastapi==0.113.0
uvicorn==0.29.0
pydantic==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20